+++ tests/test_date_time_field.py
@@ -1,18 +1,11 @@
 from datetime import datetime
 from wtforms_components import DateTimeField, DateRange
-from wtforms_test import FormTestCase
-from wtforms import Form
 
-from tests import MultiDict
+from tests import MultiDict, FieldTestCase
 
 
-class TestDateTimeField(FormTestCase):
-    def init_form(self, **kwargs):
-        class TestForm(Form):
-            test_field = DateTimeField(**kwargs)
-
-        self.form_class = TestForm
-        return self.form_class
+class TestDateTimeField(FieldTestCase):
+    field_class = DateTimeField
 
     def test_assigns_min_and_max(self):
         form_class = self.init_form(