+++ tests/test_date_time_local_field.py
@@ -1,18 +1,10 @@
 from datetime import datetime
 from wtforms_components import DateTimeLocalField, DateRange
-from wtforms_test import FormTestCase
-from wtforms import Form
-
-from tests import MultiDict
+from tests import MultiDict, FieldTestCase
 
 
-class TestDateTimeLocalField(FormTestCase):
-    def init_form(self, **kwargs):
-        class TestForm(Form):
-            test_field = DateTimeLocalField(**kwargs)
-
-        self.form_class = TestForm
-        return self.form_class
+class TestDateTimeLocalField(FieldTestCase):
+    field_class = DateTimeLocalField
 
     def test_assigns_min_and_max(self):
         form_class = self.init_form(